Refactored the directory structure and the associated files.. TODO tests
