+++ .gitignore
@@ -172,3 +172,7 @@ gradle-app.setting
 **/build/
 
 # End of https://www.gitignore.io/api/node,macos,gradle,windows,visualstudiocode
+.classpath
+.project
+.settings/org.eclipse.buildship.core.prefs
+package-lock.json
