+++ config/default.json
@@ -1,0 +1,21 @@
+{
+  "logger" : {
+    "level" : "debug"
+  },
+  "port" : 80,
+  "ssl": {
+    "enabled" : false,
+    "port" : 443,
+    "cert" : "/ssl/cert.pem",
+    "key" : "/ssl/key.pem"
+  },
+  "marklogic": {
+    "connection": {
+      "host":     "localhost",
+      "port":     8096,
+      "user":     "test-geo-data-services-reader",
+      "password": "test-geo-data-services-reader",
+      "authType": "DIGEST"
+    }
+  }
+}
