+++ src/koop/index.js
@@ -1,14 +1,14 @@
 /*
- * Copyright © 2017 MarkLogic Corporation
+ * Copyright © 2019 MarkLogic Corporation
  */
 
-const pkg = require('./package.json')
+const pkg = require('../../package.json');
 const provider = {
   name: 'marklogic',
   hosts: false,
   Model: require('./marklogic'),
   version: pkg.version,
   type: 'provider'
-}
+};
 
-module.exports = provider
+module.exports = provider;
