+++ test/gradle-test.properties
@@ -3,9 +3,9 @@ mlAppName=esri-connector
 # MarkLogic Geo Data Services settings
 # Consult your MarkLogic configuration and sure they match
 mlHost=localhost
-mlRestPort=8095
-koopMlUsername=geo-data-services-reader
-koopMlPassword=geo-data-services-reader
+mlRestPort=8096
+koopMlUsername=test-geo-data-services-reader
+koopMlPassword=test-geo-data-services-reader
 
 # Koop server setttings
 # These are the ports the feature service will be available on
