+++ test/gradle.properties
@@ -1,0 +1,19 @@
+mlAppName=esri-connector
+
+# MarkLogic Geo Data Services settings
+# Consult your MarkLogic configuration and sure they match
+mlHost=localhost
+mlRestPort=8096
+koopMlUsername=geo-data-services-reader
+koopMlPassword=geo-data-services-reader
+
+# Koop server setttings
+# These are the ports the feature service will be available on
+koopHost=localhost
+koopPort=80
+koopSSLEnabled=false
+koopSSLPort=443
+koopSSLCert=ssl/cert.pem
+koopSSLKey=ssl/key.pem
+
+serviceDescriptorPermissions=admin,update,admin,read
